+++ hutool-core/src/test/java/com/xiaoleilu/hutool/bean/BeanUtilTest.java
@@ -1,4 +1,4 @@
-package com.xiaoleilu.hutool.util;
+package com.xiaoleilu.hutool.bean;
 
 import java.lang.reflect.Type;
 import java.util.HashMap;